♻️ Speech2text: Refactor for docker deployment
